requirements.txt
beautifulsoup4==4.12.3
lxml==5.2.1
PyYAML==6.0.2
requests==2.32.3
